+++ src/main/java/com/spotify/github/jackson/PrettyPrinter.java
@@ -1,0 +1,41 @@
+/*-
+ * -\-\-
+ * github-api
+ * --
+ * Copyright (C) 2016 - 2020 Spotify AB
+ * --
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * -/-/-
+ */
+package com.spotify.github.jackson;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+/**
+ * Utility class capable of pretty printing objects to json.
+ *
+ * @author Marko Kanala, Meliora Ltd (marko.kanala@meliora.fi)
+ */
+public class PrettyPrinter {
+    protected static Json prettyPrintJson = Json.createPrettyPrinter();
+
+    private PrettyPrinter() {};
+
+    public static String toString(final Object o) {
+        try {
+            return prettyPrintJson.toJson(o);
+        } catch (JsonProcessingException je) {
+            return "toString failed with " + je.getMessage() + " for: " + o;
+        }
+    }
+}