+++ CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to this fork will be documented in this file.
 - remove shading: makes OkHttpClient configurable
 - fix Integer id() declarations "Numeric value out of range of int" JsonMappingExceptions
 - REST: Extend IssueClient to include Issue related stuff
+- impl pretty printing toString method for immutables